Integro con papa

+++ src/main/java/com/example/demo/model/pojos/SaldoInicialCentroDeCostos.java
@@ -26,7 +26,7 @@ public class SaldoInicialCentroDeCostos implements Serializable{
 	private IdSaldoInCCostos elId;
 	
 	@Column(precision = 2)
-	private float debito;
+	private double debito;
 	@Column(precision = 2)
-	private float credito;
+	private double credito;
 }
